checkIsMobile & checkIsMobileViewPort moved to utils
