+++ src/app/common/StickyHeader.tsx
@@ -5,8 +5,8 @@ import { Select } from '../../exports';
 import { getCombinedClass } from '../../utils/getCombinedClass';
 import { MOBILE_ONLY_LOGO_AND_TITLE_ID, THEME_OPTIONS } from '../constants';
 import { MIXPANEL_EVENT_PROPERTIES, MIXPANEL_EVENTS } from '../constants/mixpanel';
-import { checkIsMobile } from '../helpers';
 import { getCurrentScreen } from '../routes/routesConfig';
+import { checkIsMobile, checkIsMobileViewPort } from '../utils';
 import { mixpanel } from '../utils/mixpanel';
 
 function windowResizeEventListenerEffect(stickyBelowLogoContainer: RefObject<HTMLDivElement>) {
@@ -14,12 +14,13 @@ function windowResizeEventListenerEffect(stickyBelowLogoContainer: RefObject<HTM
 		return;
 	}
 
+	const isMobileViewPort = checkIsMobileViewPort();
 	const isMobile = checkIsMobile();
 
 	const logoContainer = document.querySelector(`#${MOBILE_ONLY_LOGO_AND_TITLE_ID}`);
 	const { height } = logoContainer?.getBoundingClientRect() ?? {};
 
-	if (Boolean(height) && isMobile) {
+	if (Boolean(height) && ((isMobile ?? false) || isMobileViewPort)) {
 		stickyBelowLogoContainer.current.style.top = `${height}px`;
 	} else {
 		stickyBelowLogoContainer.current.style.top = `0px`;
