--- src/app/helpers/index.ts
@@ -1,5 +1,0 @@
-import { MOBILE_BREAKPOINT } from '../constants';
-
-export function checkIsMobile(){
-	return document.body.clientWidth <= MOBILE_BREAKPOINT;
-}
